requirements.txt
crewai==0.102.0
gradio==5.20.0
openpyxl==3.1.5
python-dotenv==1.0.1
langchain-groq==0.1.5
